+++ assets/CHANGELOG.md
@@ -4,6 +4,7 @@
 - feat(android): Android TV support
 - feat: add troubleshoot page
 - feat(windows): left click on tray icon opens app
+- feat(linux): add Control+Q shortcut to exit app
 - fix: also scan multicast when pressing on a subnet sync button
 - fix(android): missing app icon on Android 7
 - fix(android,ios): show error message when saving to gallery failed